Fixed wrong type in definition

+++ src/PUBLIC_INCLUDES/rmn/ezscint.h
@@ -169,7 +169,7 @@ int32_t c_ezget_subgridids(int32_t gdid, int32_t *subgrid);
 int32_t f77name(ezget_subgridids)(int32_t *gdid, int32_t *subgrid);
 
 void f77name(ez_calcdist)(float *distance, float *lat1, float *lon1, float *lat2, float *lon2);
-void f77name(ez_calcdist2)(float *distance, float *lat1, float *lon1, float *lat2, float *lon2);
+void f77name(ez_calcdist2)(double *distance, float *lat1, float *lon1, float *lat2, float *lon2);
 void f77name(c_ez_calcarea_rect)(float *area, float lat1, float lon1, float lat2, float lon2);
 void f77name(ez_calcarea)(float *area, float lats[], float lons[]);
 void f77name(ez_calcarea2)(float *area, float lats[], float lons[]);
